renamed staking to simplestake
