+++ x/simplestake/errors.go
