+++ x/simplestake/types.go
